Refactor unflaten_index loop to use size_t for index iteration; add LazyEval and LazyOps for tensor expression evaluation and operations

+++ include/tensr/util/IndexUtils.hpp
@@ -28,7 +28,7 @@ std::vector<size_t> unflaten_index(const size_t idx, std::vector<size_t>& shape)
     std::vector<size_t> unflat_idx(shape.size());
     std::vector<size_t> stride = compute_strides(shape);
     size_t idx_rem = idx;
-    for (int i = 0; i < shape.size(); i++) {
+    for (size_t i = 0; i < shape.size(); i++) {
         unflat_idx[i] = idx_rem / stride[i];
         idx_rem %= stride[i];
     }
